fix:changes id datatypes for users

+++ event-planner/api/demo/src/main/java/com/api/demo/services/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User not found"));
     }
